requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gspread==3.1.0
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dotenv==0.10.3
python-http-client==3.1.0
requests==2.22.0
rsa==4.0
sendgrid==6.0.5
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
